Changed name of spec file for open-feedback-modal and reorganised tests using nested describe calls.
